Only show tables the DB user can access
We make a lot of use of per-user schemas and have a lot of tables that
users can't access, so listing all schemas/tables makes the Data
Explorer fairly unfriendly to use. Let's read the information schema,
which respects the permissions of the database connection, to only show
tables that the connection can access.