Vendor dependencies to serve locally
We want to be able to deploy data-explorer, a derived project, in
containers that don't have access to the Internet, so we need to make
sure any dependencies are served locally rather than from CloudFlare.

It looks like we don't need some of these dependencies so I've commented
them out in order to not have network errors/hit our CSP. But in case we
do need them at some point in the future, they're there and easy to find
out what else we might need to vendor.